+++ blue-tile-engine-lite/engine/rpg/Skill.h
@@ -10,6 +10,6 @@ public:
     Skill(const char * name);
     ~Skill();
     std::string m_name;
-    virtual int activate(Stats & stats, Character & target);
+    virtual float activate(Stats & stats, Character & target);
     std::string getName();
 };
